+++ src/core/EVM/parseEVMStepErrors.unit.spec.ts
@@ -5,25 +5,26 @@ import {
 import { beforeAll, describe, expect, it } from 'vitest'
 import { buildStepObject } from '../../../tests/fixtures.js'
 import { setupTestEnvironment } from '../../../tests/setup.js'
-import { LiFiErrorCode } from '../../utils/errors.js'
-import { parseError } from './parseError.js'
+import { LiFiErrorCode } from '../../utils/errors/constants.js'
+import type { LiFiSDKError } from '../../utils/errors/SDKError.js'
+import { parseEVMStepErrors } from './parseEVMStepErrors.js'
 
 beforeAll(setupTestEnvironment)
 
-describe('parseError', () => {
+describe('parseEVMStepErrors', () => {
   describe('parseWalletError', () => {
     describe('when the error does not contain a code', () => {
       it('should return an UnknownError with the default message if no message is set', async () => {
-        const parsedError = await parseError('Oops')
+        const parsedError = await parseEVMStepErrors(new Error('Oops'))
 
         expect(parsedError.message).toEqual('Unknown error occurred.')
         expect(parsedError.code).toEqual(LiFiErrorCode.InternalError)
       })
 
       it('should return an UnknownError with the given message', async () => {
-        const parsedError = await parseError({
-          message: 'Somethings fishy',
-        })
+        const parsedError = await parseEVMStepErrors(
+          new Error('Somethings fishy')
+        )
 
         expect(parsedError.message).toEqual('Somethings fishy')
         expect(parsedError.code).toEqual(LiFiErrorCode.InternalError)
@@ -32,10 +33,9 @@ describe('parseError', () => {
 
     describe('when the error contains an unknown error code', () => {
       it('should return an UnknownError', async () => {
-        const parsedError = await parseError({
-          code: 1337,
-          message: 'Somethings fishy',
-        })
+        const error = new Error('Somethings fishy')
+        ;(error as any).code = 1337
+        const parsedError = await parseEVMStepErrors(error)
 
         expect(parsedError.message).toEqual('Somethings fishy')
         expect(parsedError.code).toEqual(LiFiErrorCode.InternalError)
@@ -45,10 +45,10 @@ describe('parseError', () => {
     // TODO: revisit
     describe.skip('when the error contains a rpc error code', () => {
       it('should return a RPCError with the metamask error message', async () => {
-        const parsedError = await parseError({
-          code: MetaMaskErrorCodes.rpc.methodNotFound,
-          message: 'Somethings fishy',
-        })
+        const error = new Error('Somethings fishy')
+        ;(error as any).code = MetaMaskErrorCodes.rpc.methodNotFound
+
+        const parsedError = await parseEVMStepErrors(error)
 
         expect(parsedError.message).toEqual(
           getMessageFromCode(MetaMaskErrorCodes.rpc.methodNotFound)
@@ -57,41 +57,39 @@ describe('parseError', () => {
       })
 
       it('should return a RPCError with a custom message if underpriced', async () => {
-        const parsedError = await parseError({
-          code: MetaMaskErrorCodes.rpc.internal,
-          message: 'RPC called failed: transaction underpriced',
-        })
+        const error = new Error('RPC called failed: transaction underpriced')
+        ;(error as any).code = MetaMaskErrorCodes.rpc.internal
+
+        const parsedError = await parseEVMStepErrors(error)
 
         expect(parsedError.message).toEqual('Transaction is underpriced.')
         expect(parsedError.code).toEqual(LiFiErrorCode.TransactionUnderpriced)
       })
     })
-
+    // TODO: review tests
     describe('when the error contains a provider error code', () => {
-      it('should return a ProviderError with the metamask error message', async () => {
-        const parsedError = await parseError({
-          code: MetaMaskErrorCodes.provider.unsupportedMethod,
-          message: 'Somethings fishy',
-        })
+      it.only('should return a ProviderError with the metamask error message', async () => {
+        const error = new Error('Somethings fishy')
+        ;(error as any).code = MetaMaskErrorCodes.provider.unsupportedMethod
 
-        expect(parsedError.message).toEqual(
-          getMessageFromCode(MetaMaskErrorCodes.provider.unsupportedMethod)
+        const parsedError = await parseEVMStepErrors(error)
+
+        expect((parsedError as LiFiSDKError).cause.message).toEqual(
+          'Somethings fishy'
         )
-        expect(parsedError.code).toEqual(
-          MetaMaskErrorCodes.provider.unsupportedMethod
-        )
+        expect(parsedError.code).toEqual(LiFiErrorCode.InternalError)
       })
     })
 
     // TODO: revisit
     describe.skip('when no step is passed to the parser', () => {
       it('should return a default htmlMessage', async () => {
-        const parsedError = await parseError({
-          code: MetaMaskErrorCodes.rpc.methodNotFound,
-          message: 'Somethings fishy',
-        })
+        const error = new Error('Somethings fishy')
+        ;(error as any).code = MetaMaskErrorCodes.provider.unsupportedMethod
 
-        expect(parsedError.htmlMessage).toEqual(
+        const parsedError = await parseEVMStepErrors(error)
+
+        expect(parsedError.cause.htmlMessage).toEqual(
           // eslint-disable-next-line max-len
           "Transaction was not sent, your funds are still in your wallet, please retry.<br/>If it still doesn't work, it is safe to delete this transfer and start a new one."
         )
@@ -101,15 +99,12 @@ describe('parseError', () => {
     // TODO: revisit
     describe.skip('when a step is passed to the parser', () => {
       it('should include the token information in the htmlMessage', async () => {
-        const parsedError = await parseError(
-          {
-            code: MetaMaskErrorCodes.rpc.methodNotFound,
-            message: 'Somethings fishy',
-          },
-          buildStepObject({})
-        )
+        const error = new Error('Somethings fishy')
+        ;(error as any).code = MetaMaskErrorCodes.rpc.methodNotFound
 
-        expect(parsedError.htmlMessage).toEqual(
+        const parsedError = await parseEVMStepErrors(error, buildStepObject({}))
+
+        expect(parsedError.cause.htmlMessage).toEqual(
           // eslint-disable-next-line max-len
           "Transaction was not sent, your funds are still in your wallet (1.5 USDC on Polygon), please retry.<br/>If it still doesn't work, it is safe to delete this transfer and start a new one."
         )
@@ -119,17 +114,16 @@ describe('parseError', () => {
     // TODO: revisit
     describe.skip('when a process is passed to the parser', () => {
       it('should include the explorer link in the htmlMessage', async () => {
+        const error = new Error('Somethings fishy')
+        ;(error as any).code = MetaMaskErrorCodes.rpc.methodNotFound
         const step = buildStepObject({ includingExecution: true })
-        const parsedError = await parseError(
-          {
-            code: MetaMaskErrorCodes.rpc.methodNotFound,
-            message: 'Somethings fishy',
-          },
+        const parsedError = await parseEVMStepErrors(
+          error,
           step,
           step.execution?.process[0]
         )
 
-        expect(parsedError.htmlMessage).toEqual(
+        expect(parsedError.cause.htmlMessage).toEqual(
           // eslint-disable-next-line max-len
           'Transaction was not sent, your funds are still in your wallet (1.5 USDC on Polygon), please retry.<br/>If it still doesn\'t work, it is safe to delete this transfer and start a new one.<br>You can check the failed transaction&nbsp;<a href="https://example.com" target="_blank" rel="nofollow noreferrer">here</a>.'
         )
