+++ src/index.ts
@@ -26,5 +26,5 @@ export * from './services/api.js'
 export * from './services/balance.js'
 export * from './services/getNameServiceAddress.js'
 export * from './types/index.js'
-export * from './utils/errors.js'
-export { LiFiBaseError, type ErrorCode } from './utils/errors.js'
+export * from './utils/errors/index.js'
+export { type ErrorCode } from './utils/errors/index.js'
